Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int n1,n2,res;
+    double n1,n2,res;
     char op;
-    scanf("%d %d %c",&n1,&n2,&op);
+    scanf("%lf %lf %c",&n1,&n2,&op);
     if(op=='+'){
         res=n1+n2;
         printf("%d",res);
